add jwt to several feature

+++ app/Controllers/Api/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Controllers\Api;
 use CodeIgniter\RESTful\ResourceController;
 use App\Models\Category;
 use CodeIgniter\HTTP\RequestInterface;
+use Firebase\JWT\JWT;
 
 class CategoryController extends ResourceController
 {
@@ -57,88 +58,35 @@ class CategoryController extends ResourceController
      *
      * @return mixed
      */
-    public function create()
-    {
-        $model = new Category();
+    public function create() {
+        $key = getenv('TOKEN_SECRET');
+        $header = $this->request->getServer('HTTP_AUTHORIZATION');
+        if (!$header) return $this->failUnauthorized('Akses token diperlukan');
+        $token = explode(' ', $header)[1];
 
-        $rules = [
-            'name' => 'required',
-        ];
-
-        $messages = [
-            "name" => [
-                "required" => "{field} tidak boleh kosong",
-            ],
-        ];
-
-        $response;
-        if($this->validate($rules, $messages)) {
-            $data = [
-              'name' => $this->request->getVar('name'),
+        try {
+		    $decoded = JWT::decode($token, $key, ['HS256']);
+            $model = new Category();
+            $rules = [
+                'name' => 'required',
             ];
-
-            $model->insert($data);
-            $response = [
-                'status'   => 201,
-                'success'    => 201,
-                'messages' => [
-                    'success' => 'Kategori berhasil dibuat'
-                ]
+            $messages = [
+                "name" => [
+                    "required" => "{field} tidak boleh kosong",
+                ],
             ];
-        }else{
-            $response = [
-                'status'   => 400,
-                'error'    => 400,
-                'messages' => $this->validator->getErrors(),
-            ];
-        }
-
-
-        return $this->respondCreated($response);  
-    }
-
-    /**
-     * Return the editable properties of a resource object
-     *
-     * @return mixed
-     */
-    public function edit($id = null)
-    {
-        //
-    }
-
-    /**
-     * Add or update a model resource, from "posted" properties
-     *
-     * @return mixed
-     */
-    public function update($id = null)
-    {
-        $model = new Category();
-
-        $rules = [
-            'name' => 'required',
-        ];
-
-        $messages = [
-            "name" => [
-                "required" => "{field} tidak boleh kosong",
-            ],
-        ];
-
-        $response;
-        if($model->find($id)){
+            $response;
             if($this->validate($rules, $messages)) {
                 $data = [
-                  'name' => $this->request->getRawInput('name'),
+                  'name' => $this->request->getVar('name'),
                 ];
-
-                $model->update($id, $data['name']);
+    
+                $model->insert($data);
                 $response = [
                     'status'   => 201,
                     'success'    => 201,
                     'messages' => [
-                        'success' => 'Kategori berhasil di perbarui'
+                        'success' => 'Kategori berhasil dibuat'
                     ]
                 ];
             }else{
@@ -148,16 +96,70 @@ class CategoryController extends ResourceController
                     'messages' => $this->validator->getErrors(),
                 ];
             }
-        }else{
-            $response = [
-                'status'   => 400,
-                'error'    => 400,
-                'messages' => 'Data tidak ditemukan',
-            ];
+            return $this->respondCreated($response);  
+        } catch (\Throwable $th) {
+            return $this->fail('Akses token tidak sesuai');
         }
+    }
+
+    public function edit($id = null)
+    {
+        //
+    }
 
 
-        return $this->respondCreated($response);
+    public function update($id = null) {
+        $key = getenv('TOKEN_SECRET');
+        $header = $this->request->getServer('HTTP_AUTHORIZATION');
+        if (!$header) return $this->failUnauthorized('Akses token diperlukan');
+        $token = explode(' ', $header)[1];
+
+        try {
+		    $decoded = JWT::decode($token, $key, ['HS256']);
+            $model = new Category();
+            $rules = [
+                'name' => 'required',
+            ];
+    
+            $messages = [
+                "name" => [
+                    "required" => "{field} tidak boleh kosong",
+                ],
+            ];
+    
+            $response;
+            if($model->find($id)){
+                if($this->validate($rules, $messages)) {
+                    $data = [
+                      'name' => $this->request->getRawInput('name'),
+                    ];
+    
+                    $model->update($id, $data['name']);
+                    $response = [
+                        'status'   => 201,
+                        'success'    => 201,
+                        'messages' => [
+                            'success' => 'Kategori berhasil di perbarui'
+                        ]
+                    ];
+                }else{
+                    $response = [
+                        'status'   => 400,
+                        'error'    => 400,
+                        'messages' => $this->validator->getErrors(),
+                    ];
+                }
+            }else{
+                $response = [
+                    'status'   => 400,
+                    'error'    => 400,
+                    'messages' => 'Data tidak ditemukan',
+                ];
+            }
+            return $this->respondCreated($response);
+        } catch (\Throwable $th) {
+            return $this->fail('Akses token tidak sesuai');
+        }
     }
 
     /**
@@ -165,22 +167,30 @@ class CategoryController extends ResourceController
      *
      * @return mixed
      */
-    public function delete($id = null)
-    {
-        $model = new Category();
+    public function delete($id = null) {
+        $key = getenv('TOKEN_SECRET');
+        $header = $this->request->getServer('HTTP_AUTHORIZATION');
+        if (!$header) return $this->failUnauthorized('Akses token diperlukan');
+        $token = explode(' ', $header)[1];
 
-        if($model->find($id)){
-            $model->delete($id);
-            $response = [
-                'status'   => 200,
-                'success'    => 200,
-                'messages' => [
-                    'success' => 'Kategori berhasil di hapus'
-                ]
-            ];
-            return $this->respondDeleted($response);
-        }else{
-            return $this->failNotFound('Data tidak di temukan');
+        try {
+		    $decoded = JWT::decode($token, $key, ['HS256']);
+            $model = new Category();
+            if($model->find($id)){
+                $model->delete($id);
+                $response = [
+                    'status'   => 200,
+                    'success'    => 200,
+                    'messages' => [
+                        'success' => 'Kategori berhasil di hapus'
+                    ]
+                ];
+                return $this->respondDeleted($response);
+            }else{
+                return $this->failNotFound('Data tidak di temukan');
+            }
+        } catch (\Throwable $th) {
+            return $this->fail('Akses token tidak sesuai');
         }
     }
 }
